# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ChangeEventCreate": {
        "description": "Compact change metadata accepted from authenticated delivery tooling.",
        "properties": {
          "actor": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "change_type": {
            "enum": [
              "deployment",
              "configuration",
              "feature_flag"
            ],
            "title": "Change Type",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "revision": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "service_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Service Name",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "summary": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "service_name",
          "change_type",
          "summary"
        ],
        "title": "ChangeEventCreate",
        "type": "object"
      },
      "ChangeEventResponse": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "change_type": {
            "enum": [
              "deployment",
              "configuration",
              "feature_flag"
            ],
            "title": "Change Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "service_name": {
            "title": "Service Name",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "id",
          "project_id",
          "service_name",
          "change_type",
          "summary",
          "occurred_at",
          "created_at"
        ],
        "title": "ChangeEventResponse",
        "type": "object"
      },
      "CustomerInvestigationResponse": {
        "description": "Deliberately narrow response contract for embedded customer-facing UIs.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "diagnosed",
              "insufficient_evidence"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "status",
          "answer"
        ],
        "title": "CustomerInvestigationResponse",
        "type": "object"
      },
...
      "EvidenceGraph": {
        "description": "Engineer-only explanation graph built from an existing investigation.",
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "entry_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Node Id"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "root_cause": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RootCause"
              },
              {
                "type": "null"
              }
            ]
          },
          "root_cause_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause Node Id"
          },
          "status": {
            "enum": [
              "diagnosed",
              "insufficient_evidence"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "status",
          "root_cause",
          "nodes",
          "edges",
          "recommended_actions"
        ],
        "title": "EvidenceGraph",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvestigationRequest": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "exception_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exception Message"
          },
          "exception_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exception Type"
          },
          "http_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Status"
          },
          "metric_samples": {
            "items": {
              "$ref": "#/components/schemas/MetricSample"
            },
            "title": "Metric Samples",
            "type": "array"
          },
          "page": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "service_calls": {
            "items": {
              "$ref": "#/components/schemas/ServiceCallSample"
            },
            "title": "Service Calls",
            "type": "array"
          },
          "span_samples": {
            "items": {
              "$ref": "#/components/schemas/SpanSample"
            },
            "title": "Span Samples",
            "type": "array"
          },
          "trace_duration_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Duration Ms"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "question"
        ],
        "title": "InvestigationRequest",
        "type": "object"
      },
      "InvestigationResponse": {
        "description": "Internal/engineer investigation result. Never expose this whole model to customers.",
        "properties": {
          "customer_answer": {
            "title": "Customer Answer",
            "type": "string"
          },
          "engineer_answer": {
            "title": "Engineer Answer",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "root_cause": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RootCause"
              },
              {
                "type": "null"
              }
            ]
          },
          "service_topology": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceTopology"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "diagnosed",
              "insufficient_evidence"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "status",
          "root_cause",
          "evidence",
          "customer_answer",
          "engineer_answer",
          "recommended_actions"
        ],
        "title": "InvestigationResponse",
        "type": "object"
      },
...
      "InvestigatorTraceRequest": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "page": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "trace_id": {
            "maxLength": 64,
            "minLength": 16,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "question"
        ],
        "title": "InvestigatorTraceRequest",
        "type": "object"
      },
...
      "TraceInvestigationRequest": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "page": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "question": {
            "default": "Why did my last operation fail?",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 64,
            "minLength": 16,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id"
        ],
        "title": "TraceInvestigationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Evidence-first AI production support engineer.",
    "title": "ProdMind",
    "version": "0.9.0-rc.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/changes": {
      "post": {
        "description": "Record compact deployment/config metadata from trusted delivery tooling.",
        "operationId": "record_change_api_v1_changes_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ProdMind-Engineer-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Engineer-Key"
            }
          },
          {
            "in": "header",
            "name": "X-ProdMind-Project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Change"
      }
    },
    "/api/v1/investigate": {
      "post": {
        "operationId": "investigate_failure_api_v1_investigate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ProdMind-Engineer-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Engineer-Key"
            }
          },
          {
            "in": "header",
            "name": "X-ProdMind-Project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Investigate Failure"
      }
    },
    "/api/v1/investigate/trace": {
      "post": {
        "operationId": "investigate_trace_api_v1_investigate_trace_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ProdMind-Engineer-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Engineer-Key"
            }
          },
          {
            "in": "header",
            "name": "X-ProdMind-Project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Investigate Trace"
      }
    },
    "/api/v1/investigate/trace/graph": {
      "post": {
        "operationId": "investigate_trace_graph_api_v1_investigate_trace_graph_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ProdMind-Engineer-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Engineer-Key"
            }
          },
          {
            "in": "header",
            "name": "X-ProdMind-Project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceGraph"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Investigate Trace Graph"
      }
    },
    "/api/v1/investigator/trace": {
      "post": {
        "description": "Explain a current trace investigation without allowing AI to replace RCA.",
        "operationId": "investigate_trace_with_ai_api_v1_investigator_trace_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ProdMind-Engineer-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Engineer-Key"
            }
          },
          {
            "in": "header",
            "name": "X-ProdMind-Project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigatorTraceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIInvestigatorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Investigate Trace With Ai"
      }
    },
    "/api/v1/support": {
      "post": {
        "operationId": "support_failure_api_v1_support_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ProdMind-Project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerInvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Support Failure"
      }
    },
    "/api/v1/support/trace": {
      "post": {
        "operationId": "support_trace_api_v1_support_trace_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ProdMind-Project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Prodmind-Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerInvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Support Trace"
      }
    },
...
  }
}